Updates according to fltk-1.1.x
git-svn-id: file:///fltk/svn/branches/branch-1.2@3882 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_Tile.cxx
@@ -1,5 +1,5 @@
 //
-// "$Id: Fl_Tile.cxx,v 1.5.2.5.2.3.2.5 2004/07/27 16:04:26 easysw Exp $"
+// "$Id: Fl_Tile.cxx,v 1.5.2.5.2.3.2.6 2004/11/09 01:52:33 rokan Exp $"
 //
 // Tile widget for the Fast Light Tool Kit (FLTK).
 //
@@ -37,7 +37,7 @@
 
 void Fl_Tile::position(int oix, int oiy, int newx, int newy) {
   Fl_Widget*const* a = array();
-  int* p = sizes();
+  short* p = sizes();
   p += 8; // skip group & resizable's saved size
   for (int i=children(); i--; p += 4) {
     Fl_Widget* o = *a++;
@@ -69,7 +69,7 @@ void Fl_Tile::resize(int X,int Y,int W,int H) {
   int dy = Y-y();
   int dw = W-w();
   int dh = H-h();
-  int* p = sizes();
+  short* p = sizes();
   // resize this (skip the Fl_Group resize):
   Fl_Widget::resize(X,Y,W,H);
   // find bottom-right of resiable:
@@ -132,8 +132,8 @@ int Fl_Tile::handle(int event) {
     int oldx = 0;
     int oldy = 0;
     Fl_Widget*const* a = array();
-    int* q = sizes();
-    int* p = q+8;
+    short* q = sizes();
+    short* p = q+8;
     for (int i=children(); i--; p += 4) {
       Fl_Widget* o = *a++;
       if (o == resizable()) continue;
@@ -197,5 +197,5 @@ int Fl_Tile::handle(int event) {
 }
 
 //
-// End of "$Id: Fl_Tile.cxx,v 1.5.2.5.2.3.2.5 2004/07/27 16:04:26 easysw Exp $".
+// End of "$Id: Fl_Tile.cxx,v 1.5.2.5.2.3.2.6 2004/11/09 01:52:33 rokan Exp $".
 //
